requirements.txt
Flask==2.3.2
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.5
